+++ src/Client/UI/Variant.ts
@@ -6,25 +6,34 @@ export const Variant = Object.freeze({
 	/**
 	 * A dark variant.
 	 */
-	Dark: "dark",
+	Dark: "Dark",
 
 	/**
 	 * A light variant.
 	 */
-	Light: "light",
+	Light: "Light",
 
 	/**
 	 * A primary variant.
 	 */
-	Primary: "primary",
+	Primary: "Primary",
 
 	/**
 	 * A secondary variant.
 	 */
-	Secondary: "secondary"
+	Secondary: "Secondary"
 });
 
 /**
  * Defines tone variants.
  */
 export type Variant = typeof Variant[keyof typeof Variant];
+
+/**
+ * Returns the CSS representation of the specified variant.
+ * @param variant The variant.
+ * @returns The CSS representation of the specified variant.
+ */
+export function toCss(variant: Variant): string {
+	return variant.toLowerCase();
+}