+++ src/Server/UI/Variant.cs
@@ -35,9 +35,9 @@ public enum Variant {
 public static class VariantExtensions {
 
 	/// <summary>
-	/// Returns the Bootstrap representation of the specified variant.
+	/// Returns the CSS representation of the specified variant.
 	/// </summary>
 	/// <param name="variant">The variant.</param>
-	/// <returns>The Bootstrap representation of the specified variant.</returns>
-	public static string ToBoostrap(this Variant variant) => variant.ToString().ToLowerInvariant();
+	/// <returns>The CSS representation of the specified variant.</returns>
+	public static string ToCss(this Variant variant) => variant.ToString().ToLowerInvariant();
 }